+++ src/trader/strategy/geuant_strategy.hpp
@@ -99,11 +99,11 @@ public:
         const PriceInfo& price_info,
         bool do_liquidate,
         int32_t position_in_lots,
-        const std::map<uint32_t, OutstandingOrder>& outstanding_orders,
+        const std::map<uint64_t, OutstandingOrder>& outstanding_orders,
         std::map<int64_t, int32_t>& bid_place_orders,
         std::map<int64_t, int32_t>& ask_place_orders,
-        std::vector<uint32_t>& bid_cancel_orders,
-        std::vector<uint32_t>& ask_cancel_orders
+        std::vector<uint64_t>& bid_cancel_orders,
+        std::vector<uint64_t>& ask_cancel_orders
     ) override;
 
 private:
@@ -116,11 +116,11 @@ private:
     int32_t get_order_qty_in_lots(int64_t price_in_min_ticks, int32_t qty_limit_in_lots);
 
     void choose_orders_to_cancel(
-        const std::map<uint32_t, OutstandingOrder>& outstanding_orders,
+        const std::map<uint64_t, OutstandingOrder>& outstanding_orders,
         std::map<int64_t, int32_t>& bid_place_orders,
         std::map<int64_t, int32_t>& ask_place_orders,
-        std::vector<uint32_t>& bid_cancel_orders,
-        std::vector<uint32_t>& ask_cancel_orders
+        std::vector<uint64_t>& bid_cancel_orders,
+        std::vector<uint64_t>& ask_cancel_orders
     );
 };
